Dashboard Summary services added

+++ Models/Server/ServerDisk.cs
@@ -7,7 +7,7 @@ namespace DSM.UI.Api.Models.Server
 {
     public class ServerDisk
     {
-        public long DiskId { get; set; }
+        public int DiskId { get; set; }
         public string DiskName { get; set; }
         public int? DiskCapacity { get; set; }
         public int? DiskFreeSpace { get; set; }
